Updates for MongoDB.Driver 2.0 support
