--- NLog.MongoDB.Tests/App.config
@@ -1,9 +1,0 @@
-<?xml version="1.0"?>
-<configuration>
-	<connectionStrings>
-		<add name="MongoDB" connectionString="mongodb://mongo:db@ds035607.mongolab.com:35607/nlog"/>
-	</connectionStrings>
-	<startup>
-		<supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5"/>
-	</startup>
-</configuration>
