--- NLog.MongoDB.Tests/NLog.config
@@ -1,66 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<nlog xmlns="http://www.nlog-project.org/schemas/NLog.xsd"
-      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-			autoReload="true">
-	<extensions>
-		<add assembly="NLog.MongoDB" />
-	</extensions>
-	
-  <targets>
-    
-		<target xsi:type="MongoDB" name="Mongo"
-				database="nlog"
-        host="ds035607.mongolab.com"
-				port="35607"
-        username="mongo"
-				password="db" />
-
-		<target xsi:type="MongoDB" name="Mongo-ConnectionString"
-        connectionString="mongodb://mongo:db@ds035607.mongolab.com:35607/nlog" />
-
-		<target xsi:type="MongoDB" name="Mongo-ConnectionName"
-        connectionName="MongoDB" />
-
-		<target xsi:type="MongoDB" name="Mongo-ConnectionName-Fields"
-        connectionName="MongoDB">
-			<field name="timestamp" layout="${date}" />
-			<field name="level" layout="${level}" />
-			<field name="message" layout="${message}" />
-			<field name="exception" layout="${exception}" />
-		</target>
-
-		<target xsi:type="MongoDB" name="Mongo-ConnectionName-TypedFields"
-				connectionName="MongoDB">
-			<field name="timestamp" />
-			<field name="level" />
-			<field name="message" />
-			<field name="exception" />
-			<field name="transactionId" />
-		</target>
-
-		<target xsi:type="MongoDB" name="Mongo-CappedWithId"
-        connectionName="MongoDB"
-        useCappedCollection="true"
-				cappedCollectionSize="10000"
-        createIdField="true">
-			<field name="timestamp" />
-		</target>
-
-		<target xsi:type="MongoDB" name="Mongo-CappedWithoutId"
-        connectionName="MongoDB"
-        useCappedCollection="true"
-				cappedCollectionSize="10000">
-			<field name="timestamp" />
-		</target>
-
-	</targets>
-	<rules>
-		<logger name="testMongo"  minlevel="Info" writeTo="Mongo"/>
-		<logger name="testMongoConnectionString" minlevel="Info" writeTo="Mongo-ConnectionString"/>
-		<logger name="testMongoConnectionName" minlevel="Info" writeTo="Mongo-ConnectionName"/>
-		<logger name="testDynamicFields" minlevel="Info" writeTo="Mongo-ConnectionName-Fields"/>
-		<logger name="testDynamicTypedFields" minlevel="Info" writeTo="Mongo-ConnectionName-TypedFields"/>
-		<logger name="cappedWithId" minlevel="Info" writeTo="Mongo-CappedWithId"/>
-		<logger name="cappedWithoutId" minlevel="Info" writeTo="Mongo-CappedWithoutId"/>
-	</rules>
-</nlog>
