--- NLog.MongoDB.Tests/packages.config
@@ -1,9 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="FluentAssertions" version="2.1.0.0" targetFramework="net45" />
-  <package id="mongocsharpdriver" version="1.8.3" targetFramework="net45" />
-  <package id="Moq" version="4.1.1309.1617" targetFramework="net45" />
-  <package id="NLog" version="2.1.0" targetFramework="net45" />
-  <package id="NLog.Config" version="2.1.0" targetFramework="net45" />
-  <package id="NLog.Schema" version="2.1.0" targetFramework="net45" />
-</packages>
