--- NLog.MongoDB.sln
@@ -1,33 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 2012
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NLog.MongoDB", "NLog.MongoDB\NLog.MongoDB.csproj", "{D2081421-EE28-4C13-B8BD-584423DEE08A}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NLog.MongoDB.Tests", "NLog.MongoDB.Tests\NLog.MongoDB.Tests.csproj", "{2B3F7590-CCF1-472E-B356-0F7BB17D35E2}"
-EndProject
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = ".nuget", ".nuget", "{390B707E-75F0-4492-A972-5EDC639E1EAB}"
-	ProjectSection(SolutionItems) = preProject
-		.nuget\NuGet.Config = .nuget\NuGet.Config
-		.nuget\NuGet.exe = .nuget\NuGet.exe
-		.nuget\NuGet.targets = .nuget\NuGet.targets
-	EndProjectSection
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{D2081421-EE28-4C13-B8BD-584423DEE08A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{D2081421-EE28-4C13-B8BD-584423DEE08A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{D2081421-EE28-4C13-B8BD-584423DEE08A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{D2081421-EE28-4C13-B8BD-584423DEE08A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2B3F7590-CCF1-472E-B356-0F7BB17D35E2}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2B3F7590-CCF1-472E-B356-0F7BB17D35E2}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2B3F7590-CCF1-472E-B356-0F7BB17D35E2}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2B3F7590-CCF1-472E-B356-0F7BB17D35E2}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-EndGlobal
