--- NLog.MongoDB/FieldSearchResult.cs
@@ -1,8 +1,0 @@
-﻿namespace NLog.MongoDB
-{
-	public sealed class FieldSearchResult
-	{
-		public bool Succeded { get; set; }
-		public object Value { get; set; }
-	}
-}
