--- NLog.MongoDB/IRepository.cs
@@ -1,12 +1,0 @@
-using System;
-using MongoDB.Bson;
-
-namespace NLog.MongoDB
-{
-	public interface IRepository : IDisposable
-	{
-		void Insert(string collectionName, BsonDocument item);
-	    
-		void CheckCollection(string collectionName, long collectionSize, long? collectionMaxItems, bool createIdField);
-	}
-}
