--- NLog.MongoDB/IRepositoryProvider.cs
@@ -1,9 +1,0 @@
-using MongoDB.Driver;
-
-namespace NLog.MongoDB
-{
-	public interface IRepositoryProvider
-	{
-        IRepository GetRepository(MongoServerSettings settings, string databaseName);
-	}
-}
