--- NLog.MongoDB/MongoServerProvider.cs
@@ -1,12 +1,0 @@
-using MongoDB.Driver;
-
-namespace NLog.MongoDB
-{
-	public class MongoServerProvider : IRepositoryProvider
-	{
-        public IRepository GetRepository(MongoServerSettings settings, string databaseName)
-	    {
-	        return new MongoRepository(settings, databaseName);
-	    }
-	}
-}
