--- NLog.MongoDB/NLog.MongoDB.nuspec
@@ -1,19 +1,0 @@
-<?xml version="1.0"?>
-<package >
-  <metadata>
-    <id>NLog.MongoDB</id>
-    <version>0.7.0</version>
-    <authors>Jason Wicker,Tim Kellogg</authors>
-    <owners>Jason Wicker</owners>
-    <licenseUrl>https://github.com/Logrythmik/NLog.MongoDB/blob/master/License.txt</licenseUrl>
-    <projectUrl>http://logrythmik.github.com/NLog.MongoDB/</projectUrl>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>An NLog Target for MongoDB persistence. This package targets Mongo C# Driver 1.8</description>
-    <copyright>Copyright 2013</copyright>
-    <tags>logging nlog mongodb driver mongo log</tags>
-    <dependencies>
-      <dependency id="mongocsharpdriver" version="1.8.3" />
-      <dependency id="nlog" version="2.1.0" />
-    </dependencies>
-  </metadata>
-</package>
