--- NLog.MongoDB/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="mongocsharpdriver" version="1.8.3" targetFramework="net45" />
-  <package id="NLog" version="2.1.0" targetFramework="net45" />
-</packages>
