+++ NLog.MongoDB20.Tests/App.config
@@ -1,0 +1,9 @@
+<?xml version="1.0"?>
+<configuration>
+	<connectionStrings>
+		<add name="MongoDB" connectionString="mongodb://localhost:27017/nlog"/>
+	</connectionStrings>
+	<startup>
+		<supportedRuntime version="v4.0" sku=".NETFramework,Version=v4.5.1"/>
+	</startup>
+</configuration>
