+++ NLog.MongoDB20.Tests/NLog.config
@@ -1,0 +1,13 @@
+<?xml version="1.0" encoding="utf-8" ?>
+<nlog xmlns="http://www.nlog-project.org/schemas/NLog.xsd"
+      xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance">
+  <extensions>
+    <add assembly="NLog.MongoDB20"/>
+  </extensions>
+  <targets>
+    <target name="Mongo" type="MongoDB20" connectionString="mongodb://localhost:27017/nlog" />
+  </targets>
+  <rules>
+    <logger name="*" minLevel="Info" appendTo="Mongo"/>
+  </rules>
+</nlog>
