+++ NLog.MongoDB20.Tests/packages.config
@@ -1,0 +1,11 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="FluentAssertions" version="3.3.0" targetFramework="net451" />
+  <package id="MongoDB.Bson" version="2.0.0" targetFramework="net451" />
+  <package id="MongoDB.Driver" version="2.0.0" targetFramework="net451" />
+  <package id="MongoDB.Driver.Core" version="2.0.0" targetFramework="net451" />
+  <package id="Moq" version="4.2.1502.0911" targetFramework="net451" />
+  <package id="NLog" version="3.2.1" targetFramework="net451" />
+  <package id="NLog.Config" version="3.2.1" targetFramework="net451" />
+  <package id="NLog.Schema" version="3.2.1" targetFramework="net451" />
+</packages>
