+++ NLog.MongoDB20/FieldSearchResult.cs
@@ -1,0 +1,8 @@
+﻿namespace NLog.MongoDB20
+{
+	public sealed class FieldSearchResult
+	{
+		public bool Succeded { get; set; }
+		public object Value { get; set; }
+	}
+}
