+++ NLog.MongoDB20/IRepository.cs
@@ -1,0 +1,12 @@
+using System;
+using MongoDB.Bson;
+
+namespace NLog.MongoDB20
+{
+	public interface IRepository : IDisposable
+	{
+		void Insert(string collectionName, BsonDocument item);
+	    
+		void CheckCollection(string collectionName, long collectionSize, long? collectionMaxItems, bool createIdField);
+	}
+}
