+++ NLog.MongoDB20/IRepositoryProvider.cs
@@ -1,0 +1,9 @@
+using MongoDB.Driver;
+
+namespace NLog.MongoDB20
+{
+	public interface IRepositoryProvider
+	{
+        IRepository GetRepository(MongoClientSettings settings, string databaseName);
+	}
+}
