+++ NLog.MongoDB20/MongoDBTargetField.cs
@@ -1,0 +1,22 @@
+﻿using NLog.Config;
+using NLog.Layouts;
+
+namespace NLog.MongoDB20
+{
+	[NLogConfigurationItem]
+	public sealed class MongoDBTargetField
+	{
+		public MongoDBTargetField() { }
+
+		public MongoDBTargetField(string name, Layout layout)
+		{
+			Name = name;
+			Layout = layout;
+		}
+
+		[RequiredParameter]
+		public string Name { get; set; }
+
+		public Layout Layout { get; set; }
+	}
+}
