+++ NLog.MongoDB20/packages.config
@@ -1,0 +1,8 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="CreateNewNuGetPackageFromProjectAfterEachBuild" version="1.8.5" targetFramework="net451" />
+  <package id="MongoDB.Bson" version="2.0.0" targetFramework="net451" />
+  <package id="MongoDB.Driver" version="2.0.0" targetFramework="net451" />
+  <package id="MongoDB.Driver.Core" version="2.0.0" targetFramework="net451" />
+  <package id="NLog" version="3.2.1" targetFramework="net451" />
+</packages>
